+++ FluxViewer.DataAccess/Export/Exporters/CsvFileExporter.cs
@@ -35,7 +35,7 @@ public class CsvFileExporter : FileExporter
         }
     }
 
-    public override long CalculateApproximateExportSizeInBytes(int numOfPoint)
+    public override long CalculateApproximateExportSizeInBytes(long numOfPoint)
     {
         long numOfBytesInOneElement = 0;
         // Всё берём по максимуму
